+++ css/style.css
@@ -85,15 +85,12 @@ section > article {
   height: 100%;
 }
 
-.column-add-btn:hover {
-  color: #3E95E3;
-}
-
 .column-delete-btn:hover {
   color: #FF4D4D;
 }
 
-#column-add-btn {
+#fab-btn {
+  transition: 0.5s;
   position: fixed;
   right: 5%;
   bottom: 5%;
@@ -106,6 +103,9 @@ section > article {
   color: #fff;
   border-radius: 50%;
   background-color: #0175DE;
+}
+#fab-btn:hover {
+  box-shadow: 0 0.5vh 1vh 0 #3E95E3;
 }
 
 #go-column-add-modal-btn {
@@ -325,7 +325,7 @@ header > i {
   background-color: #F0FAE9;
   border-bottom: 1px dotted black;
 }
-.log-frame:last-child {
+.log-frame :last-child {
   border-bottom: none;
 }
 .log-frame > .log-emotion-area {
