+++ index.html
@@ -29,18 +29,10 @@
     </header>
     <!-- main -->
     <main>
-        <button id="go-card-clear-btn">
-            <span class="material-symbols-outlined">
-                delete
-            </span>
-        </button>
-        <button id="go-column-add-modal-btn">
-            <span class="material-symbols-outlined">add_circle</span>
-        </button>
-        <button id="go-search-btn">
-            <span class="material-symbols-outlined">search</span>
-        </button>
-        <button id="column-add-btn">+</button>
+        <button id="go-card-clear-btn"><span class="material-symbols-outlined">delete</span></button>
+        <button id="go-column-add-modal-btn"><span class="material-symbols-outlined">add_circle</span></button>
+        <button id="go-search-btn"><span class="material-symbols-outlined">search</span></button>
+        <button id="fab-btn">+</button>
         <!-- menu -->
         <aside id="menu">
             <section id="menu-header">
