+++ js/common/commonFunction.js
@@ -1,9 +1,11 @@
-import { statusList, statusNameList, JSON_DATA } from "../json_data/json_data.js";
-import { mainTag, addEventToFabBtn } from "../component/column.js";
-import { columnTemplate, cardTemplate } from "../templates/template.js";
+
+import { 
+    addEventToFabBtn, addEventToMenuBtns, addEventToModalButtons, addEventToColumnAddButton 
+} from "../component/button.js";
+import { mainTag } from "../component/column.js";
 import { makeCardDragEvent } from "../drag/addDragEvent.js";
-import { addEventToMenuBtns } from "../component/menu.js";
-import { addEventToModalButtons, addEventToFabButton } from "../component/modal.js";
+import { statusList, statusNameList, JSON_DATA } from "../json_data/json_data.js";
+import { columnTemplate, cardTemplate } from "../templates/template.js";
 
 // (target)의 css (key)를 (value)로 바꿉니다.
 function changeCSS(targetNode, key, value) {
@@ -34,10 +36,10 @@ function loadInitialData() {
 
 // 웹 페이지에 필요한 이벤트를 추가합니다.
 function addEventsToWebsite() {
+    loadInitialData();
     addEventToMenuBtns();
     addEventToModalButtons();
-    addEventToFabButton();
-    loadInitialData();
+    addEventToColumnAddButton();
     addEventToFabBtn();
 }
