+++ js/drag/dragCard.js
@@ -1,6 +1,6 @@
-import { dragIDManager } from "./dragIDManager.js";
 import { addChildAfterParent } from "../common/commonFunction.js";
 import { makeShadedNode, makeLightNode } from "./dragEffect.js";
+import { dragIDManager } from "./dragIDManager.js";
 
 // 드래그 중인 카드를 기록합니다.
 function dragCard(event) {
