+++ js/json_data/json_data.js
@@ -1,7 +1,5 @@
-
-import { findCardContent } from "../component/card.js"
-import { findCardTitle } from "../component/card.js"
-import { updateColumnLength } from "../component/column.js"
+import { findCardContent, findCardTitle } from "../component/card.js";
+import { updateColumnLength } from "../component/column.js";
 
 // 새로운 상태값이 입력되면? statusList의 길이 값을 대입해주면 된다!
 const TODO = 0
